requirements.txt
# pandas
# numpy
# sklearn

actionable-recourse
Flask>=2.0.2
flask-cors>=3.0.10
gunicorn==20.1.0

numpy>=1.14.0
scipy>=0.19.1
scikit-learn>=0.19.1
pandas>=0.23.4
cplex>=12.8
mip==1.12.0
prettytable>=0.7.2
matplotlib>=3.0.2
seaborn>=0.9.0
traitlets>=4.3.2
IPython>=7.2.0
pytest>=4.3.0
tqdm>=4.24.0
jinja2>=2.10.1